agent/requirements.txt
fastapi
uvicorn
jupyterlab
notebook
python-multipart
openai
python-dotenv 